--- a/gemm/Cargo.toml
+++ b/gemm/Cargo.toml
@@ -37,6 +37,7 @@ rand = "0.8.5"
 nalgebra = "0.32.2"
 assert_approx_eq = "1.1.0"
 rayon = "1.7"
+num_cpus = "1.16.0"
 
 [[bench]]
 name = "bench"